requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
decorator==5.1.1
defusedxml==0.7.1
Django==5.0
django-allauth==0.59.0
django-bootstrap4==24.1
django-filter==23.5
django-tables2==2.7.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
drf-spectacular-sidecar==2023.12.1
gunicorn==21.2.0
h11==0.14.0
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
packaging==23.2
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.16.2
self==2020.12.3
soupsieve==2.5
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.27.1
